requirements.txt
langchain-openai
fastapi
uvicorn
python-dotenv
pydantic
pytest
httpx
supabase
tqdm
rich
